refactor(ui): Move ui/(cli|gui)/PdfAutoA11y(CLI|GUI) to ui/(Cli|Gui)
