+++ CONTRIBUTING.md
@@ -13,8 +13,8 @@ This project is organized around a sequential pipeline that:
 
 Most code lives under `src/main/java/net/boyechko/pdf/autoa11y/`:
 
-- `ui/`: CLI/GUI entry points and reporting (e.g. `PdfAutoA11yCLI`,
-  `PdfAutoA11yGUI`, `FormattedListener`, `LoggingListener`)
+- `ui/`: CLI/GUI entry points and reporting (e.g. `Cli`, `Gui`,
+  `FormattedListener`, `LoggingListener`)
 - `core/`: orchestration (e.g. `ProcessingService`, `ProcessingDefaults`)
 - `document/`: PDF IO and context helpers (e.g. `PdfCustodian`,
   `DocContext`)
