+++ pdf-autoa11y-gui
@@ -8,4 +8,4 @@ if [ "$SCRIPT_DIR/src/main/java" -nt "$SCRIPT_DIR/target/classes" ] || [ ! -d "$
 fi
 
 java -cp "$SCRIPT_DIR/target/classes:$(cd "$SCRIPT_DIR" && mvn -q dependency:build-classpath -Dmdep.outputFile=/dev/stdout)" \
-    net.boyechko.pdf.autoa11y.ui.gui.PdfAutoA11yGUI "$@"
+    net.boyechko.pdf.autoa11y.ui.gui.Gui "$@"
